\name{swedenr-package}
\alias{swedenr-package}
\alias{swedenr}
\docType{package}
\title{
R-functions specific to Swedish conditions.
}
\description{
R-functions specific to Swedish conditions maintained by members of Stockholm R useR group.

The package GitHub homepage:
https://github.com/SthlmR/swedenr

Stockholm R useR group page:
http://www.meetup.com/StockholmR/
}
\details{
\tabular{ll}{
Package: \tab swedenr\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-09-13\cr
License: \tab GNU General Public License, version 2\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Måns Magnusson

Maintainer: Måns Magnusson <mans.magnusson@liu.se>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
